Move hashers to separate class
